requirements.txt
pytz==2024.1
schedule==1.2.0
python-dotenv==1.0.0
psutil==5.9.6
